test: co-locate queries.ts test coverage in queries.test.ts
Move unit tests for isSqlParam, isSyncedTable, validateRecord, and
buildQuery from connector.test.ts to queries.test.ts so the shared
module has co-located coverage per project convention (foo.ts →
foo.test.ts). Unique test cases (ON CONFLICT SET exclusions, EXCLUDED
references, NOW() timestamp handling, junction table scoping) are
preserved; duplicates already covered in queries.test.ts are dropped.
connector.test.ts now contains only createNeonConnector integration
tests.

Closes #90
