+++ src/sync/connector.test.ts
@@ -1,287 +1,10 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
 
-vi.mock("@powersync/web", () => ({
-  UpdateType: { PUT: "PUT", PATCH: "PATCH", DELETE: "DELETE" },
-}));
-
 vi.mock("./events", () => ({
   dispatchSyncError: vi.fn(),
 }));
 
-import { UpdateType } from "@powersync/web";
 import { dispatchSyncError } from "./events";
-import {
-  buildQuery,
-  isSqlParam,
-  isSyncedTable,
-  validateRecord,
-} from "./queries";
-
-describe("isSqlParam", () => {
-  it("returns true for null", () => {
-    expect(isSqlParam(null)).toBe(true);
-  });
-
-  it("returns true for string", () => {
-    expect(isSqlParam("hello")).toBe(true);
-  });
-
-  it("returns true for number", () => {
-    expect(isSqlParam(42)).toBe(true);
-  });
-
-  it("returns true for boolean", () => {
-    expect(isSqlParam(true)).toBe(true);
-    expect(isSqlParam(false)).toBe(true);
-  });
-
-  it("returns false for undefined", () => {
-    expect(isSqlParam(undefined)).toBe(false);
-  });
-
-  it("returns false for object", () => {
-    expect(isSqlParam({ key: "value" })).toBe(false);
-  });
-
-  it("returns false for array", () => {
-    expect(isSqlParam([1, 2])).toBe(false);
-  });
-});
-
-describe("isSyncedTable", () => {
-  it("returns true for a synced table", () => {
-    expect(isSyncedTable("tricks")).toBe(true);
-  });
-
-  it("returns false for a non-synced table", () => {
-    expect(isSyncedTable("users")).toBe(false);
-  });
-});
-
-describe("validateRecord", () => {
-  it("accepts a valid table and columns", () => {
-    expect(() =>
-      validateRecord("tricks", { id: "1", name: "Ambitious Card" })
-    ).not.toThrow();
-  });
-
-  it("throws for a disallowed column", () => {
-    expect(() =>
-      validateRecord("tricks", { id: "1", not_a_column: "bad" })
-    ).toThrow('Disallowed column "not_a_column" on table "tricks"');
-  });
-});
-
-describe("buildQuery", () => {
-  describe("PUT", () => {
-    it("generates an upsert query", () => {
-      const result = buildQuery(
-        UpdateType.PUT,
-        "tricks",
-        "abc",
-        {
-          id: "abc",
-          name: "Card Trick",
-        },
-        "user-1"
-      );
-
-      expect(result.query).toContain("INSERT INTO");
-      expect(result.query).toContain("ON CONFLICT");
-      expect(result.query).toContain('"tricks"');
-      expect(result.params).toEqual(["abc", "Card Trick", "user-1"]);
-    });
-
-    it("excludes id from ON CONFLICT SET clause", () => {
-      const result = buildQuery(
-        UpdateType.PUT,
-        "tricks",
-        "abc",
-        {
-          id: "abc",
-          name: "Card Trick",
-        },
-        "user-1"
-      );
-
-      const onConflictPart = result.query.split("DO UPDATE SET")[1];
-      expect(onConflictPart).not.toContain('"id"');
-    });
-
-    it("uses EXCLUDED references in ON CONFLICT SET clause", () => {
-      const result = buildQuery(
-        UpdateType.PUT,
-        "tricks",
-        "abc",
-        {
-          id: "abc",
-          name: "Card Trick",
-        },
-        "user-1"
-      );
-
-      expect(result.query).toContain("EXCLUDED");
-    });
-
-    it("adds deleted_at = NULL resurrection clause when deleted_at not in record", () => {
-      const result = buildQuery(
-        UpdateType.PUT,
-        "tricks",
-        "abc",
-        {
-          id: "abc",
-          user_id: "user1",
-          name: "Test",
-        },
-        "user1"
-      );
-
-      const onConflictPart = result.query.split("DO UPDATE SET")[1];
-      expect(onConflictPart).toContain('"deleted_at" = NULL');
-    });
-
-    it("uses NOW() for updated_at in ON CONFLICT SET clause", () => {
-      const result = buildQuery(
-        UpdateType.PUT,
-        "tricks",
-        "abc",
-        {
-          id: "abc",
-          name: "Card Trick",
-          updated_at: "2024-01-01T00:00:00Z",
-        },
-        "user-1"
-      );
-
-      const onConflictPart = result.query.split("DO UPDATE SET")[1];
-      expect(onConflictPart).toContain('"updated_at" = NOW()');
-      expect(onConflictPart).not.toContain('EXCLUDED."updated_at"');
-    });
-  });
-
-  describe("PATCH", () => {
-    it("generates an update query excluding id from SET", () => {
-      const result = buildQuery(
-        UpdateType.PATCH,
-        "tricks",
-        "abc",
-        {
-          id: "abc",
-          name: "Updated Trick",
-        },
-        "user-1"
-      );
-
-      expect(result.query).toContain("UPDATE");
-      expect(result.query).toContain("SET");
-      expect(result.query).toContain("WHERE");
-      expect(result.query).not.toContain('SET "id"');
-      expect(result.params).toEqual(["Updated Trick", "abc", "user-1"]);
-    });
-
-    it("appends NOW() for updated_at instead of using client value", () => {
-      const result = buildQuery(
-        UpdateType.PATCH,
-        "tricks",
-        "abc",
-        {
-          id: "abc",
-          name: "Updated Trick",
-          updated_at: "2024-01-01T00:00:00Z",
-        },
-        "user-1"
-      );
-
-      expect(result.query).toContain('"updated_at" = NOW()');
-      // updated_at client value should not be in params
-      expect(result.params).not.toContain("2024-01-01T00:00:00Z");
-      expect(result.params).toEqual(["Updated Trick", "abc", "user-1"]);
-    });
-  });
-
-  describe("DELETE", () => {
-    it("generates a soft-delete query using server-side NOW()", () => {
-      const result = buildQuery(
-        UpdateType.DELETE,
-        "tricks",
-        "abc",
-        {},
-        "user-1"
-      );
-
-      expect(result.query).toContain('"deleted_at" = NOW()');
-      expect(result.query).toContain('"updated_at" = NOW()');
-      expect(result.query).toContain('WHERE "id" = $1');
-      expect(result.params).toEqual(["abc", "user-1"]);
-    });
-  });
-});
-
-describe("buildQuery with userId scoping", () => {
-  it("adds user_id WHERE clause to PUT when userId is provided", () => {
-    const result = buildQuery(
-      UpdateType.PUT,
-      "tricks",
-      "abc",
-      {
-        id: "abc",
-        user_id: "user-1",
-        name: "Card Trick",
-      },
-      "user-1"
-    );
-    expect(result.query).toContain('"user_id"');
-    expect(result.params).toContain("user-1");
-  });
-
-  it("adds user_id WHERE clause to PATCH when userId is provided", () => {
-    const result = buildQuery(
-      UpdateType.PATCH,
-      "tricks",
-      "abc",
-      {
-        id: "abc",
-        name: "Updated Trick",
-      },
-      "user-1"
-    );
-    expect(result.query).toContain('"user_id"');
-    expect(result.params).toContain("user-1");
-  });
-
-  it("adds user_id WHERE clause to DELETE when userId is provided", () => {
-    const result = buildQuery(
-      UpdateType.DELETE,
-      "tricks",
-      "abc",
-      {
-        id: "abc",
-      },
-      "user-1"
-    );
-    expect(result.query).toContain('"user_id"');
-    expect(result.params).toContain("user-1");
-  });
-
-  it("includes user_id clause for junction tables", () => {
-    const result = buildQuery(
-      UpdateType.PUT,
-      "routine_tricks",
-      "abc",
-      {
-        id: "abc",
-        user_id: "user-1",
-        routine_id: "r1",
-        trick_id: "t1",
-        position: 1,
-      },
-      "user-1"
-    );
-    // Junction tables now have user_id in SYNCED_COLUMNS
-    expect(result.params).toContain("user-1");
-    expect(result.query).toContain('"user_id"');
-  });
-});
 
 describe("createNeonConnector", () => {
   function importWithEnv(
